+++ Changelog.md
@@ -2,7 +2,12 @@
 
 ## 0.2.0 - (2025-07-13)
 
-
+- Big refactor of the app
+- Added keyboard controls
+- Added nice check for CLOC and error message if not installed
+- Added grouping by language
+- Added grouping by file type
+- Added option for fullscreen or inline mode in CLI (-f flag)
 
 ## 0.1.0 - (2025-07-12)